tweak minis (more)
the current rules are that every mini is always enterable from any
entrance, but the only active exit in _any_ mini is the one the player
last used.

which is lame; it makes the original puzzle impossible to
make. but the obvious solution (every exit is always active and acts as
a cloning device if there are multiple minis) will also make the solution not work because it clones the player

maybe clone everything besides the player? i need to think about this
more

oh also the current implementation is broken because crates push
onto the levelStack