interesting (more)
this version of the puzzle has the tutorial at the _end_,
after you thought you'd won. that will probably feel unfair...
maybe just use that for the bonus? but the whole point of this
level is to teach the incerator method...

also, the recorded cheeses (cheese and cheese2)
don't work but i fear they're dangerously close