fix bug (more)
newpush: 11121121112211000002223333

newpush was being weird when you did the dup1.lvl thing
inside green - it pushed purple against the wall, which was weird.

but that was b/c the player was unpushable

now, it tries to push the player backwards but fails b/c the player
already has a frameStack, so the purple box gets culled.

it.... makes enough sense for me to leave it for now.
but please don't design any puzzles around this

wait shoot should the player get cloned by dup1???? prolly not...?
if not, should they end up dead if they recursed too far first?
or, always come out of the last one? probably that.