wip froot sorta works?
it's a level, and it requires doing the looping stuff.

it doesn't teach it too well though

also, how am i gonna make advanced levels with this?? its v hard to
make looping necessary

maybe itd be easier to make levels where looping is prohibited
but easy to do accidentally?